Refactor test folder
